fix(hash): Fix hash slab having duplicate keys
If you called `HashMemorySlab::intern_path` with a cache that was not
used to create it, then it would insert duplicate keys into the hash
slab.

This was causing problems with has lookup and making the logs basically
unreadable while I was debugging, but I was more concerned about trying
to replace a file that looked up the wrong components.

This was fixed by setting a flag inside of the `HashMemorySlab` that
indicates whether or not it has already been "finalized". If it's
already been finalized, then you cannot have used the same cache that
you finalizd it with (since calling `finalize` consumes the cache).
Instead we now also search the slab's buckets for the hash as well.

This also changed the API for creating a cache to ensure that we can't
accidentally create a default cache for a slab which was already
finalized once.