feat: Implement unsharing for fighter source slots
Stage shared files are still a little bit buggy, which means
that there is work yet to be done. The current bug is that battlefield
files which are likely being unshared are acting as if they are still
shared if (and only if) it is played before small battlefield. If I play
small battlefield but not normal battlefield, then small battlefield
looks normal. But once I've played on battlefield once, small
battlefield never looks the same again.

There's also like a 35% chance that you just infinitely hang after
advancing past the title screen.