feat: Split loading and decompression
This commit splits the loading of compressed zip files from the
decompression. When a file is loaded via a single-file
request, the file data is loaded in ResLoadingThread. If it is
compressed, it will continue to be stay compressed and will
be decompressed by ResInflateThread.

This helps divide the workflow between the two threads when loading
many consecutive files that are modded, such as when working with UI